+++ CHANGELOG.md
@@ -62,12 +62,16 @@
 - Selection manipulation\Advanced sort...: new harmonic mixing entry to apply a semi-random pattern.
 - Playlist manipulation\Harmonic mix: new harmonic mixing entry. See above.
 - Added integrity checks to global user settings files, found at '[FOOBAR PROFILE FOLDER]\js_data\presets\global\[...].json'. In particular queries are now check to ensure they are valid and will throw a popup at init otherwise. Other settings are check to ensure they contain valid values too.
+- Toolbar: new settings for text (or icon) position: left, right (default), top and bottom. These settings mimic CUI options at the buttons bar.
+- Toolbar: new settings for text scale (now independent of button and icon scale).
+- Toolbar: new settings for icon scale (now independent of button and text scale).
 ### Changed
 - Helpers: updated helpers.
 - Code cleanup.
 ### Removed
 ### Fixed
 - Other tools\Import track list: error setting custom path.
+- Toolbar: buttons' size not restored back to normal height after disabling 'Full size buttons' without reloading the panel.
 - Minor fixes.
 
 ## [3.7.0] - 2023-12-17